add support for flatmesh feature
Add support for the `Unwrap Mesh` and `Unwrap Face` features in the Mesh
Workbench.

These features allow to unfold a mesh, as detailled here :
    - https://wiki.freecadweb.org/Mesh_Workbench
    - https://www.youtube.com/watch?v=-KjMZkLfE1A

These features are enabled via the `BUILD_FLAT_MESH` compilation option
(which defaults to `OFF`), and makes use of the `pybind11` library.

This commit also adds the `pybind11` module to the flatpak app.