Update libspnav and pass along its socket
This will allow the use of space navigator input devices in FreeCAD,
verified with my Magellan Plus.

As /var/run is a symlink to /run on modern systems, this will not
require any change in spacenavd configuration either, and it won't have
any difference to users without such an input device - as it'll simply
ignore the passthrough if the socket does not exist when starting the
application.

Fixes #18